Brett adds work
Alright, so being self-conscious combined with being new to git kinda
made me want to work totally on my local machine, sorry about that. I
really wanted to have it completely working, but I’m having trouble
sending data to my server in order to save it to the DB. I tried
building it off of the model of the last lab, but then ended up
simplifying it immensely, in the JS department I mean. The script block
in scores.html contains ajax requests to communicate with php page
containing the DB (encoded as JSON). I can read that info and add it to
the html no problem (the functions getScores and updateList do this).
But the ajax call in addEntry just gives me an error, and that’s what’s
been stumping me. I’m calling addEntry on that page simply for
debugging, I imagine that it would be called in main.js at the end of a
game, with the counter variable being passed in as a parameter.